+++ src/Unreal.Core/Contracts/ITelemetryEvent.cs
@@ -7,6 +7,6 @@
     {
         public float? ReplicatedWorldTimeSeconds { get; set; }
 
-        public float? ReplicatedWorldTimeSecondsDouble { get; set; }
+        public double? ReplicatedWorldTimeSecondsDouble { get; set; }
     }
 }
